earth-summit-myphotos-backend-main/embedding-service/requirements.txt
fastapi==0.109.0
mangum==0.17.0
uvicorn==0.27.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime
opencv-python-headless==4.9.0.80
numpy==1.24.3
Pillow==10.0.1
boto3
requests
